Fix: the system default socket timeout is used to open URLs.
The system default socket timeout may be infinite, meaning that unresponsive URLs can cause the bot to block indefinitely.
This allows, for example, a denial-of-service attack with the !url command.
`util.ext_urlopen' was added with a default timeout of 12 seconds, and all affected modules were updated to use this instead of `urllib2.urlopen.`
`util.ext_url_opener' was modified to use the same default timeout (in fact, this is used to implement the above).